Bug fix: cast values to correct type after implicit convertion to int

+++ include/EDF/BitField.hpp
@@ -32,7 +32,7 @@ public:
     constexpr BitField( T initialValue = 0 ) : bits( initialValue ) {}
 
     constexpr void set( std::size_t startBit, std::size_t span, T value ) {
-        EDF_ASSERTD( value < (1 << span), "value must fit within \"span\"" );
+        EDF_ASSERTD( value < static_cast<std::size_t>(1 << span), "value must fit within \"span\"" );
         const auto mask = createMask( startBit, span );
         bits = (bits & ~mask) | ((value << startBit) & mask);
     }
